labs/lab-05/app/business-api/python/payment/pyproject.toml
[project]
name = "banking-assistant-payment-mcp"
version = "0.1.0"
description = "Python MCP server for payment business-api using fastmcp"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastmcp",
    "pydantic",
    "requests",
    "fastapi[all]==0.116.1",
    "uvicorn",
    "azure-monitor-opentelemetry",
    "azure-identity",
    "azure-keyvault-secrets",
]
